infinity/requirements.txt
python-dotenv
fastapi
mcp==1.7.0
fastapi-mcp
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
requests
httpx
pytz

